+++ common/components/Ad/index.js
@@ -1,6 +1,6 @@
 export CrazyAd from './CrazyAd';
 export DFP from './DFP';
-export getAdType from './getAdType';
+export * as getAdType from './getAdType';
 export GrabBag from './GrabBag';
 export OneAdIR from './OneAdIR';
 export OneAdICIP from './OneAdICIP';