+++ CHANGELOG.md
@@ -5,13 +5,19 @@
 而項目則基於 [Semantic Versioning](http://semver.org/lang/zh-TW/).
 
 ## 1.0.9 - 2017-07-14
+### Added
+- Popin 的 meta @esbb48
+
 ### Changed
+- 使用 imgApi @esbb48
 - 專欄分類頁的 Slide 可以控制 @esbb48
 - 更新立院直播的新聞來源 @esbb48
 
 ### Fixed
 - 地方新聞內頁都是生活廣告 @esbb48
 - 避免選到的專題沒有新聞 @esbb48
+### Fixed
+- 溫度計指標 PV 太大會消失在盡頭 @esbb48
 
 ## 1.0.8 - 2017-07-12
 ### Fixed
@@ -39,7 +45,7 @@
 ## 1.0.4 - 2017-07-06
 ### Added
 - 增加好康報報版位
-- 增加 Popoin
+- 增加 Popin
 - 加上觸發渲染twitter嵌入元素 @appleoxxo
 ### Fixed
 - OneAd、DFP 廣告版位調整 @esbb48